Regroup root API files
