--- cgo/kuzzle/disconnect.go
@@ -1,15 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#include "kuzzle.h"
-*/
-import "C"
-import (
-	"github.com/kuzzleio/sdk-go/kuzzle"
-)
-
-//export kuzzle_wrapper_disconnect
-func kuzzle_wrapper_disconnect(k *C.kuzzle) {
-	(*kuzzle.Kuzzle)(k.instance).Disconnect()
-}
