--- cgo/kuzzle/refresh_index.go
@@ -1,29 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#cgo LDFLAGS: -ljson-c
-	#include <stdlib.h>
-	#include "kuzzle.h"
-*/
-import "C"
-import (
-	"github.com/kuzzleio/sdk-go/kuzzle"
-)
-
-//export kuzzle_wrapper_refresh_index
-func kuzzle_wrapper_refresh_index(k *C.kuzzle, index *C.char, options *C.query_options) *C.shards {
-	result := (*C.shards)(C.calloc(1, C.sizeof_shards))
-	opts := SetQueryOptions(options)
-
-	shards, err := (*kuzzle.Kuzzle)(k.instance).RefreshIndex(C.GoString(index), opts)
-	if err != nil {
-		Set_shards_error(result, err)
-		return result
-	}
-
-	result.total = C.int(shards.Total)
-	result.successful = C.int(shards.Successful)
-	result.failed = C.int(shards.Failed)
-	return result
-}
