--- cgo/kuzzle/set_headers.go
@@ -1,19 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#include <kuzzle.h>
-*/
-import "C"
-import (
-	"github.com/kuzzleio/sdk-go/kuzzle"
-)
-
-//export kuzzle_wrapper_set_headers
-func kuzzle_wrapper_set_headers(k *C.kuzzle, content *C.json_object, replace C.uint) {
-	jp := JsonParser{}
-	jp.Parse(content)
-
-	r := replace != 0
-	(*kuzzle.Kuzzle)(k.instance).SetHeaders(jp.GetContent(), r)
-}
