--- cgo/kuzzle/stop_queuing.go
@@ -1,13 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#include <kuzzle.h>
-*/
-import "C"
-import "github.com/kuzzleio/sdk-go/kuzzle"
-
-//export kuzzle_wrapper_stop_queuing
-func kuzzle_wrapper_stop_queuing(k *C.kuzzle) {
-	(*kuzzle.Kuzzle)(k.instance).StopQueuing()
-}
